requirements.txt
websockets
pytest
asyncio
pytest-asyncio
pyjwt
